Add Linux port: CMake build system, platform abstraction, and emulation libraries (Phases 0-2)
Implement the first three phases of the Linux port, enabling all 12 core
emulation libraries and ATTest to compile and link on Linux with GCC/CMake.

Build system:
- Top-level CMakeLists.txt with Ninja generator, C++23, per-library targets
- AltirraCompilerFlags.cmake with GCC warning configuration
- 14 library CMakeLists.txt files (system, vdjson, ATCore, ATCPU, ATEmulation,
  ATAudio, ATIO, ATDevices, ATNetwork, ATNetworkSockets, ATDebugger, ATVM,
  Kasumi, ATTest)
- AVX2/FMA compiler flags for SIMD source files

Platform headers (#ifdef VD_PLATFORM_WINDOWS/LINUX, no Windows code modified):
- vdtypes.h: GCC/Clang compiler detection, CPU/platform macros
- atomic.h: GCC builtins alongside Clang
- int128.h: GCC AMD64 inline shift/add/sub/mul via __int128, fix shift UB at count==0
- miniwindows.h: Linux stub typedefs (VDZHWND, etc.)
- intrin.h: compatibility shim redirecting to x86intrin.h
- thread.h: Linux opaque CritSec/RWLock/CondVar storage
- file.h: VDFileHandle as int (fd) on Linux
- bitmath.h, strutil.h, math.h, memory.h, vdstdc.h: GCC builtins and POSIX equivalents

Linux platform implementations (new _linux.cpp files):
- thread_linux.cpp: pthreads (VDThread, VDCriticalSection, VDSignal, VDSemaphore, VDRWLock)
- file_linux.cpp: POSIX open/read/write/lseek64/fstat
- filesys_linux.cpp: opendir/readdir, stat, path splitting, VDParsedPath, canonical paths
- time_linux.cpp: clock_gettime monotonic, VDCallbackTimer via clock_nanosleep
- cpuaccel_linux.cpp: cpuid.h detection (SSE2 through AVX2), ARM64 getauxval
- debug_linux.cpp: fprintf assert, VDDebugPrint
- error_linux.cpp: strerror_r formatting
- date_linux.cpp: localtime_r/mktime/strftime
- registry_linux.cpp: VDRegistryProviderMemory as default provider
- nativesockets_linux.cpp: BSD sockets stub for ATNetworkSockets
- main_linux.cpp: ATTest entry point

Library #ifdef guards for GCC compatibility:
- stdafx.h files: platform-aware includes across all libraries
- Various source files: guard Windows-only headers and APIs

Test results: 19/25 pass, 5 expected failures (Windows paths, missing test data),
1 pre-existing AVX2 alignment bug in Core_FFT.

+++ src/ATIO/source/vorbisbitreader.cpp
@@ -15,6 +15,7 @@
 //	with this program. If not, see <http://www.gnu.org/licenses/>.
 
 #include "stdafx.h"
+#include <cmath>
 #include <at/atio/vorbisbitreader.h>
 
 // Read Vorbis-encoded floats. These use a different representation from IEEE floats.
@@ -38,7 +39,7 @@ float ATVorbisBitReader::ReadFloat() {
 	const float sign = v & 0x80000000U ? -1.0f : 1.0f;
 	const int exponent = std::clamp((int)((v & 0x7fe00000) >> 21) - 788, -63, 63);
 
-	return sign * ldexpf((float)mantissa, exponent);
+	return sign * std::ldexp((float)mantissa, exponent);
 }
 
 void ATVorbisBitReader::RefillSlow() {
